Convert Notifications to TS
