--- site/src/components/home/notifications/style.jsx
@@ -1,18 +1,0 @@
-const notificationsStyle = theme => ({
-    root: {
-    },
-    popup: {
-    },
-    errorTarget: {
-    },
-    errors: {
-        display: "flex",
-        flexDirection: "column"
-    },
-    errorHeader: {
-        display: "flex",
-        flexDirection: "row",
-        justifyContent: "space-between",
-        borderBottom: "solid aquamarine 2px",
-    }
-});
